+++ packages/wallet/primitives/src/network.ts
@@ -1,3 +1,8 @@
+import { Address } from 'ox'
+
+const DEFAULT_MULTICALL3_ADDRESS: Address.Address = '0xcA11bde05977b3631167028862bE2a173976CA11'
+const SEQUENCE_MULTICALL3_ADDRESS: Address.Address = '0xae96419a81516f063744206d4b5E36f3168280f8'
+
 export enum NetworkType {
   MAINNET = 'mainnet',
   TESTNET = 'testnet',
@@ -21,8 +26,11 @@ export interface Network {
     name: string
     decimals: number
   }
-  ensAddress?: string
   deprecated?: true
+  contracts?: {
+    multicall3?: Address.Address
+    ensUniversalResolver?: Address.Address
+  }
 }
 
 export const ChainId = {
@@ -130,6 +138,15 @@ export const ChainId = {
 
   // ARC
   ARC_TESTNET: 5042002,
+
+  // HYPEREVM
+  HYPEREVM: 999,
+
+  // SONIC
+  SONIC: 146,
+
+  // BERACHAIN
+  BERACHAIN: 80094,
 } as const
 
 export type ChainId = (typeof ChainId)[keyof typeof ChainId]
@@ -151,7 +168,10 @@ export const ALL: Network[] = [
       name: 'Ether',
       decimals: 18,
     },
-    ensAddress: '0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e',
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+      ensUniversalResolver: '0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e',
+    },
   },
   {
     chainId: ChainId.SEPOLIA,
@@ -168,6 +188,9 @@ export const ALL: Network[] = [
       symbol: 'sETH',
       name: 'Sepolia Ether',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -186,6 +209,9 @@ export const ALL: Network[] = [
       name: 'POL',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.POLYGON_AMOY,
@@ -202,6 +228,9 @@ export const ALL: Network[] = [
       symbol: 'aPOL',
       name: 'Amoy POL',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -220,6 +249,9 @@ export const ALL: Network[] = [
       name: 'Ether',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.BSC,
@@ -236,6 +268,9 @@ export const ALL: Network[] = [
       symbol: 'BNB',
       name: 'BNB',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -254,6 +289,9 @@ export const ALL: Network[] = [
       name: 'Testnet BNB',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.OPTIMISM,
@@ -270,6 +308,9 @@ export const ALL: Network[] = [
       symbol: 'ETH',
       name: 'Ether',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -288,6 +329,9 @@ export const ALL: Network[] = [
       name: 'Sepolia Ether',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.ARBITRUM,
@@ -304,6 +348,9 @@ export const ALL: Network[] = [
       symbol: 'ETH',
       name: 'Ether',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -322,6 +369,9 @@ export const ALL: Network[] = [
       name: 'Sepolia Ether',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.ARBITRUM_NOVA,
@@ -338,6 +388,9 @@ export const ALL: Network[] = [
       symbol: 'ETH',
       name: 'Ether',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -356,6 +409,9 @@ export const ALL: Network[] = [
       name: 'AVAX',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.AVALANCHE_TESTNET,
@@ -372,6 +428,9 @@ export const ALL: Network[] = [
       symbol: 'tAVAX',
       name: 'Testnet AVAX',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -390,6 +449,9 @@ export const ALL: Network[] = [
       name: 'XDAI',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.BASE,
@@ -406,6 +468,9 @@ export const ALL: Network[] = [
       symbol: 'ETH',
       name: 'Ether',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -424,6 +489,9 @@ export const ALL: Network[] = [
       name: 'Sepolia Ether',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.HOMEVERSE,
@@ -440,6 +508,9 @@ export const ALL: Network[] = [
       symbol: 'OAS',
       name: 'OAS',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -458,6 +529,9 @@ export const ALL: Network[] = [
       name: 'Testnet OAS',
       decimals: 18,
     },
+    contracts: {
+      multicall3: SEQUENCE_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.XAI,
@@ -474,6 +548,9 @@ export const ALL: Network[] = [
       symbol: 'XAI',
       name: 'XAI',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -492,6 +569,9 @@ export const ALL: Network[] = [
       name: 'Sepolia XAI',
       decimals: 18,
     },
+    contracts: {
+      multicall3: SEQUENCE_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.B3,
@@ -508,6 +588,9 @@ export const ALL: Network[] = [
       symbol: 'ETH',
       name: 'Ether',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -526,6 +609,9 @@ export const ALL: Network[] = [
       name: 'Ether',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.APECHAIN,
@@ -542,6 +628,9 @@ export const ALL: Network[] = [
       symbol: 'APE',
       name: 'ApeCoin',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -560,6 +649,9 @@ export const ALL: Network[] = [
       name: 'ApeCoin',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.BLAST,
@@ -576,6 +668,9 @@ export const ALL: Network[] = [
       symbol: 'ETH',
       name: 'Ether',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -594,6 +689,9 @@ export const ALL: Network[] = [
       name: 'Ether',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.TELOS,
@@ -610,6 +708,9 @@ export const ALL: Network[] = [
       symbol: 'TLOS',
       name: 'TLOS',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -628,6 +729,9 @@ export const ALL: Network[] = [
       name: 'TLOS',
       decimals: 18,
     },
+    contracts: {
+      multicall3: SEQUENCE_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.SKALE_NEBULA,
@@ -644,6 +748,9 @@ export const ALL: Network[] = [
       symbol: 'sFUEL',
       name: 'SKALE Fuel',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -662,6 +769,9 @@ export const ALL: Network[] = [
       name: 'SKALE Fuel',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.SONEIUM,
@@ -678,6 +788,9 @@ export const ALL: Network[] = [
       symbol: 'ETH',
       name: 'Ether',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -696,6 +809,9 @@ export const ALL: Network[] = [
       name: 'Ether',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.TOY_TESTNET,
@@ -712,6 +828,9 @@ export const ALL: Network[] = [
       symbol: 'TOY',
       name: 'TOY',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -730,6 +849,9 @@ export const ALL: Network[] = [
       name: 'IMX',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.IMMUTABLE_ZKEVM_TESTNET,
@@ -746,6 +868,9 @@ export const ALL: Network[] = [
       symbol: 'IMX',
       name: 'IMX',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -764,6 +889,9 @@ export const ALL: Network[] = [
       name: 'GLMR',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.MOONBASE_ALPHA,
@@ -780,6 +908,9 @@ export const ALL: Network[] = [
       symbol: 'GLMR',
       name: 'GLMR',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -798,6 +929,9 @@ export const ALL: Network[] = [
       name: 'Tez',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.ETHERLINK_SHADOWNET_TESTNET,
@@ -814,6 +948,9 @@ export const ALL: Network[] = [
       symbol: 'XTZ',
       name: 'Tez',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
   {
@@ -832,6 +969,9 @@ export const ALL: Network[] = [
       name: 'MON',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
   {
     chainId: ChainId.MONAD_TESTNET,
@@ -848,6 +988,9 @@ export const ALL: Network[] = [
       symbol: 'MON',
       name: 'MON',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
 
@@ -867,6 +1010,9 @@ export const ALL: Network[] = [
       name: 'SOMI',
       decimals: 18,
     },
+    contracts: {
+      multicall3: SEQUENCE_MULTICALL3_ADDRESS,
+    },
   },
 
   {
@@ -884,6 +1030,9 @@ export const ALL: Network[] = [
       symbol: 'STT',
       name: 'STT',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: SEQUENCE_MULTICALL3_ADDRESS,
     },
   },
 
@@ -903,6 +1052,9 @@ export const ALL: Network[] = [
       name: 'TCENT',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
 
   {
@@ -920,6 +1072,9 @@ export const ALL: Network[] = [
       symbol: 'ETH',
       name: 'ETH',
       decimals: 18,
+    },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
     },
   },
 
@@ -939,6 +1094,9 @@ export const ALL: Network[] = [
       name: 'SAND',
       decimals: 18,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
 
   {
@@ -957,7 +1115,65 @@ export const ALL: Network[] = [
       name: 'USDC',
       decimals: 6,
     },
+    contracts: {
+      multicall3: DEFAULT_MULTICALL3_ADDRESS,
+    },
   },
+
+  {
+    chainId: ChainId.HYPEREVM,
+    type: NetworkType.MAINNET,
+    name: 'hyperevm',
+    title: 'HyperEVM',
+    rpcUrl: getRpcUrl('hyperevm'),
+    logoUrl: getLogoUrl(ChainId.HYPEREVM),
+    blockExplorer: {
+      name: 'HyperEVM Explorer',
+      url: 'https://www.hyperscan.com/',
+    },
+    nativeCurrency: {
+      symbol: 'HYPE',
+      name: 'HYPE',
+      decimals: 18,
+    },
+  },
+
+  {
+    chainId: ChainId.BERACHAIN,
+    type: NetworkType.MAINNET,
+    name: 'berachain',
+    title: 'Berachain',
+    rpcUrl: getRpcUrl('berachain'),
+    logoUrl: getLogoUrl(ChainId.BERACHAIN),
+    blockExplorer: {
+      name: 'Berachain Explorer',
+      url: 'https://berascan.com/',
+    },
+    nativeCurrency: {
+      symbol: 'BEAR',
+      name: 'BEAR',
+      decimals: 18,
+    },
+  },
+
+  {
+    chainId: ChainId.SONIC,
+    type: NetworkType.MAINNET,
+    name: 'sonic',
+    title: 'Sonic',
+    rpcUrl: getRpcUrl('sonic'),
+    logoUrl: getLogoUrl(ChainId.SONIC),
+    blockExplorer: {
+      name: 'Sonic Explorer',
+      url: 'https://sonicscan.com/',
+    },
+    nativeCurrency: {
+      symbol: 'S',
+      name: 'Sonic',
+      decimals: 18,
+    },
+  },
+
 ]
 
 function getRpcUrl(networkName: string): string {